custom SpiderWebScoreView

+++ android/kreader/app/oem/Sun/src/main/java/com/onyx/android/sun/cloud/bean/StudyReportDetailBean.java
@@ -32,8 +32,8 @@ public class StudyReportDetailBean {
         public static class PointsBean {
 
             @SerializedName("class")
-            public int classX;
-            public int own;
+            public float classX;
+            public float own;
         }
     }
